fix(tga): env-expand Bitbucket app_password like token (closes #842)
PR #743 fixed `${VAR}` expansion for the Bitbucket `token` config field
but missed `app_password`.  Users who stored
`app_password = "${MY_SECRET}"` in YAML had the literal placeholder sent
as the HTTP Basic-auth password, yielding a silent 401.

Apply `crate::collect::env_expand::expand_env_var` to the config
`app_password` value before the `BITBUCKET_APP_PASSWORD` env-var
fallback, mirroring the identical treatment of `token` at lines 93-104.
Existing fallback chain semantics are preserved:
  config app_password (now env-expanded) → BITBUCKET_APP_PASSWORD → error

Also split the inline `#[cfg(test)] mod tests { … }` block out of
`client.rs` into a sibling `tests.rs` to bring `client.rs` back below
the 500-line cap (ratchets the allowlist entry down).

New tests added (all three pass):
  • app_password_env_var_expanded              — closes #842 core case
  • app_password_config_takes_precedence_over_env_fallback
  • app_password_falls_back_to_env_when_config_absent
